requirements.txt
fastapi==0.127.0
uvicorn[standard]==0.40.0
pydantic==2.12.5
python-dotenv==1.2.1
requests==2.32.5
aiohttp==3.11.11
httpx==0.28.1
APScheduler==3.11.0
python-dateutil==2.9.0